requirement.txt
torch>=2.0.0
torchvision
numpy
matplotlib
scikit-image
tqdm
pillow